requirements.txt
fastapi~=0.85.2
tensorflow==2.13.0
ultralytics==8.0.203

Pillow~=9.3.0
torchvision==0.15.2
pydicom==2.4.4
opencv-python==4.7.0.68

pandas==2.0.0